wip help: Force next-line format even if "already" too wide
TODO write tests; issue refs

In the existing logic, if one takes some option and makes its name,
plus placeholder and so on, longer and longer, then at some point
that will force the help output into next-line format in order to
avoid being excessively wide... and then as it gets longer than
that, it eventually *stops* doing so and we revert to the wide,
same-line format.

That seems backward -- making things longer should cause us to
break onto the next line, not to go back to putting the now-long
items next to each other horizontally.

Investigating where this logic came from, to try to find what the
rationale might have been: this part of the logic is ancient,
going back to when the "force next line" feature was first added,
in b7793a2f4d back in 2016-08.

I don't see in the original issue thread # 597, nor in # 587, any
discussion of this quirk.  It's not in the changelog entries added
by that commit, either.  So that doesn't turn up any subtle
surprising reason why this condition is needed after all.
Go ahead and take it out.